Controller, monitor and resource: take 3
