+++ testctrl/svc/orch/resource.go
@@ -2,6 +2,7 @@ package orch
 
 import (
 	"fmt"
+	"strings"
 	"sync"
 
 	"k8s.io/api/core/v1"
@@ -23,12 +24,10 @@ func NewResources(cs ...*types.Component) []*Resource {
 	var resources []*Resource
 
 	for _, component := range cs {
-		for i := int32(0); i < component.Replicas(); i++ {
-			resources = append(resources, &Resource{
-				name:      component.Name(),
-				component: component,
-			})
-		}
+		resources = append(resources, &Resource{
+			name:      component.Name(),
+			component: component,
+		})
 	}
 
 	return resources
@@ -46,11 +45,22 @@ func (r *Resource) Update(status v1.PodStatus) {
 	r.mux.Lock()
 	defer r.mux.Unlock()
 
+	r.podStatus = status
+
 	for _, cstatus := range status.ContainerStatuses {
 		cstate := cstatus.State
-		if cstate.Terminated != nil {
+		if cstatus.State.Terminated != nil || cstatus.LastTerminationState.Terminated != nil {
 			r.unhealthy = true
 			r.err = fmt.Errorf("resource %v: docker container has terminated, unable to recover", r)
+			return
+		}
+
+		if waitingState := cstate.Waiting; waitingState != nil {
+			if strings.Compare("CrashLoopBackOff", waitingState.Reason) == 0 {
+				r.unhealthy = true
+				r.err = fmt.Errorf("resource %v: docker container has entered crash loop", r)
+				return
+			}
 		}
 	}
 
@@ -95,88 +105,3 @@ func (r *Resource) Done() bool {
 	return r.done
 }
 
-type Monitor struct {
-	done	    bool
-	unhealthy   bool
-	resources   map[string]*Resource
-	errResource *Resource
-	mux         sync.Mutex
-}
-
-func NewMonitor() *Monitor {
-	return &Monitor{
-		resources: make(map[string]*Resource),
-	}
-}
-
-func (m *Monitor) Get(resourceName string) *Resource {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-	return m.resources[resourceName]
-}
-
-func (m *Monitor) Add(r *Resource) {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-	m.resources[r.Name()] = r
-}
-
-func (m *Monitor) Remove(resourceName string) {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-	delete(m.resources, resourceName)
-}
-
-func (m *Monitor) Update(pod *v1.Pod) error {
-	resourceName := pod.Labels["resource-name"]
-	if len(resourceName) < 1 {
-		return fmt.Errorf("monitor cannot update using pod named '%v', missing a resource-name label")
-	}
-
-	m.mux.Lock()
-	defer m.mux.Unlock()
-
-	r := m.resources[resourceName]
-	if r == nil {
-		return fmt.Errorf("monitor does not manage resource named '%v'", resourceName)
-	}
-
-	r.Update(pod.Status)
-	if r.Unhealthy() {
-		m.unhealthy = true
-		m.errResource = r
-		return r.Error()
-	}
-
-	return nil
-}
-
-func (m *Monitor) Unhealthy() bool {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-	return m.unhealthy
-}
-
-func (m *Monitor) Error() error {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-
-	if m.errResource != nil {
-		return m.errResource.Error()
-	}
-
-	return nil
-}
-
-func (m *Monitor) Done() bool {
-	m.mux.Lock()
-	defer m.mux.Unlock()
-
-	for _, r := range m.resources {
-		if !r.Done() {
-			return false
-		}
-	}
-
-	return true
-}
